--- auto-matching-nuxt/components/Card.vue.d.ts
@@ -1,5 +1,0 @@
-import { Vue } from "nuxt-property-decorator";
-import { Person } from "../types";
-export default class Card extends Vue {
-    person: Person;
-}
