+++ auto-matching-nuxt/pages/index.vue
@@ -1,40 +1,34 @@
 <template>
-  <section>
-    <h1 class="header">
-      Nuxt TypeScript Starter
-    </h1>
-    <div class="cards">
-      <Card
-        v-for="person in people"
-        :key="person.id"
-        :person="person"
-      />
-    </div>
-  </section>
+  <div id="app">
+    <HelloVue
+      val="Vue"
+      @handle-click="handleClickButton($event)"
+    />
+  </div>
 </template>
 
 <script lang="ts">
-import { Component, Vue } from 'nuxt-property-decorator';
-import { State } from 'vuex-class';
-import { Person } from '../types';
-import Card from '../components/Card.vue';
-
+import { Component, Vue } from 'vue-property-decorator';
+import HelloVue from '../components/HelloVue.vue';
 @Component({
   components: {
-    Card,
+    HelloVue,
   },
 })
-export default class extends Vue {
-  @State people: Person;
+export default class App extends Vue {
+  handleClickButton($event: string) {
+    console.log($event);
+  }
 }
 </script>
 
-<style scoped>
-.header {
-  font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
-}
-.cards {
-  display: flex;
-  flex-wrap: wrap;
+<style>
+#app {
+  font-family: "Avenir", Helvetica, Arial, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  text-align: center;
+  color: #2c3e50;
+  margin-top: 60px;
 }
 </style>
