--- auto-matching-nuxt/types/state.ts
@@ -1,6 +1,0 @@
-import { Person } from '~/types';
-
-export interface RootState {
-  people: Person[];
-}
-;
